Revert "merge master into petrovic/brush-editor-ui"
This reverts commit 682af93794833ade8f9fc572fc62bd19399d02ee, reversing
changes made to 7b4eef4659173130a7bdc1de40299b030c0b8139.

+++ libs/image/tests/kis_warp_transform_worker_test.cpp
@@ -268,8 +268,9 @@ void KisWarpTransformWorkerTest::testBackwardInterpolatorRoundTrip()
     KisFourPointInterpolatorForward f(src, dst);
     KisFourPointInterpolatorBackward b(src, dst);
 
-    for (int y = 0; y <= 100; y += 1) {
-        for (int x = 0; x <= 100; x += 1) {
+    for (qreal y = 0; y <= 100; y += 1.0) {
+        for (qreal x = 0; x <= 100; x += 1.0) {
+
             QPointF pt(x, y);
 
             QPointF fwdPt = f.map(pt);
